% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-missSBM_fit.R
\name{coef.missSBM_fit}
\alias{coef.missSBM_fit}
\title{Extract model coefficients}
\usage{
\method{coef}{missSBM_fit}(
  object,
  type = c("mixture", "connectivity", "covariates", "sampling"),
  ...
)
}
\arguments{
\item{object}{an R6 object with class [`missSBM_fit`]}

\item{type}{type of parameter that should be extracted. Either "mixture" (default), "connectivity", "covariates" or "sampling"}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A vector or matrix of coefficients extracted from the missSBM_fit model.
}
\description{
Extracts model coefficients from objects [`missSBM_fit`] returned by [estimateMissSBM()]
}
