% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampler-Class.R
\docType{data}
\name{dyadSampler}
\alias{dyadSampler}
\title{Virtual class for all dyad-centered samplers}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dyadSampler
}
\description{
Virtual class for all dyad-centered samplers
}
\keyword{datasets}
