% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_network.R
\docType{data}
\name{er_network}
\alias{er_network}
\title{ER ego centred network}
\format{A sparse symmetric matrix with 741 rows and 741 columns \code{ESR1}}
\source{
\url{https://string-db.org/}
}
\usage{
er_network
}
\description{
A dataset containing the weighted PPI network centred around the ESR1 (ER) protein
}
\examples{
data("er_network")
class(er_network)
}
\keyword{datasets}
