% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESR1.R
\docType{data}
\name{ESR1}
\alias{ESR1}
\title{ESR ego centred network}
\format{A sparse symmetric matrix with 741 rows and 741 columns \code{ESR1}}
\source{
\url{https://string-db.org/}
}
\usage{
ESR1
}
\description{
A dataset containing the weigthed PPI network centred around ESR1
}
\examples{
data(ESR1)
class(ESR1)
}
\keyword{datasets}
