% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenchblog2007.R
\docType{data}
\name{frenchblog2007}
\alias{frenchblog2007}
\title{Political Blogosphere network prior to 2007 French presidential election}
\format{An igraph object with 196 nodes. The vertex attribute "party" provides a possible
clustering of the nodes.}
\source{
\url{https://www.linkfluence.com/}
}
\usage{
frenchblog2007
}
\description{
French Political Blogosphere network dataset consists of a single day snapshot of
over 200 political blogs automatically extracted the 14 october 2006 and manually classified by
the "Observatoire Presidentielle" project. Originally part of the 'mixer' package
}
\examples{
data(frenchblog2007)
igraph::V(frenchblog2007)$party
\dontrun{
igraph::plot(frenchblog2007,
  vertex.color = factor(igraph::V(frenchblog2007)$party),
  vertex.label = NA
 )
}
}
\keyword{datasets}
