% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_missSBM.R
\name{l1_similarity}
\alias{l1_similarity}
\title{L1-similarity}
\usage{
l1_similarity(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
a vector equal to -abs(x-y)
}
\description{
Compute l1-similarity between two vectors
}
