% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingSBM_fit-Class.R
\docType{data}
\name{missingSBM_fit}
\alias{missingSBM_fit}
\title{R6 Class definition of an SBM-fit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
missingSBM_fit
}
\description{
This class is designed to adjust a Stochastic Block Model on a network with missing entries.
}
\keyword{datasets}
