% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{missSBM-defunct}
\alias{missSBM-defunct}
\alias{estimate}
\alias{sample}
\alias{simulate}
\title{Defunct Functions in Package missSBM}
\usage{
estimate(...)

sample(...)

simulate(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package missSBM
}
\details{
\code{estimate()} is replaced by \code{missSBM::estimateMissSBM}.

\code{sample()} is replaced by \code{missSBM::observeNetwork}.

\code{simualte()} is replaced by \code{sbm::sampleSimpleSBM}.
}
