% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM-package.R
\docType{package}
\name{missSBM}
\alias{missSBM}
\alias{missSBM-package}
\title{An R package for adjusting Stochastic Block Models from networks data sampled under various missing data conditions}
\description{
The SBM package provides three main functions: \code{\link{simulate}}, \code{\link{sample}}, \code{\link{estimate}}.
}
\references{
Timothée Tabouy, Pierre Barbillon & Julien Chiquet (2019) “Variational Inference for Stochastic Block Models from Sampled Data”, Journal of the American Statistical Association, <doi:10.1080/01621459.2018.1562934>
}
\author{
Timothée Tabouy, Pierre Barbillon, Julien Chiquet
}
