% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampler-Class.R
\name{networkSampler}
\alias{networkSampler}
\title{Definition of R6 Class 'networkSampling_sampler'}
\description{
This class is use to define a sampling model for a network. Inherits from 'networkSampling'.
Owns a rSampling method which takes an adjacency matrix as an input and send back an object
with class sampledNetwork.
}
\seealso{
\code{\link{sampledNetwork}}
}
\section{Super class}{
\code{\link[missSBM:networkSampling]{missSBM::networkSampling}} -> \code{networkSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{networkSampler$new()}}
\item \href{#method-rSamplingMatrix}{\code{networkSampler$rSamplingMatrix()}}
\item \href{#method-clone}{\code{networkSampler$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSampling" data-id="print">}\href{../../missSBM/html/networkSampling.html#method-print}{\code{missSBM::networkSampling$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSampling" data-id="show">}\href{../../missSBM/html/networkSampling.html#method-show}{\code{missSBM::networkSampling$show()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{networkSampler$new(type = NA, parameters = NA, nNodes = NA, directed = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rSamplingMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-rSamplingMatrix}{}}}
\subsection{Method \code{rSamplingMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{networkSampler$rSamplingMatrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{networkSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
