% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampling-Class.R
\docType{data}
\name{networkSampling}
\alias{networkSampling}
\title{definition of R6 Class 'networkSampling'}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
networkSampling
}
\description{
this virtual class is the mother of all subtypes of networkSampling (either sampler or fit)
It is used to define a sampling model for a network.
It has a rSampling method which takes an adjacency matrix as an input and send back an object with class sampledNetwork.
}
\keyword{datasets}
