% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampling_sampler-Class.R
\docType{data}
\name{networkSampling_sampler}
\alias{networkSampling_sampler}
\title{Definition of R6 Class 'networkSampling_sampler'}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
networkSampling_sampler
}
\description{
This class is use to define a sampling model for a network. Inherits from 'networkSampling'
it has a rSampling method which takes an adjacency matrix as an input and send back an object with class sampledNetwork.
}
\keyword{datasets}
