% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-networkSampler.R
\name{nodeSampler}
\alias{nodeSampler}
\title{Virtual class for all node-centered samplers}
\description{
Virtual class for all node-centered samplers

Virtual class for all node-centered samplers
}
\section{Super classes}{
\code{\link[missSBM:networkSampling]{missSBM::networkSampling}} -> \code{\link[missSBM:networkSampler]{missSBM::networkSampler}} -> \code{nodeSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-nodeSampler-clone}{\code{nodeSampler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="missSBM" data-topic="networkSampling" data-id="print"><a href='../../missSBM/html/networkSampling.html#method-networkSampling-print'><code>missSBM::networkSampling$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="missSBM" data-topic="networkSampling" data-id="show"><a href='../../missSBM/html/networkSampling.html#method-networkSampling-show'><code>missSBM::networkSampling$show()</code></a></span></li>
<li><span class="pkg-link" data-pkg="missSBM" data-topic="networkSampler" data-id="initialize"><a href='../../missSBM/html/networkSampler.html#method-networkSampler-initialize'><code>missSBM::networkSampler$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="missSBM" data-topic="networkSampler" data-id="rSamplingMatrix"><a href='../../missSBM/html/networkSampler.html#method-networkSampler-rSamplingMatrix'><code>missSBM::networkSampler$rSamplingMatrix()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-nodeSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
