% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-missSBM_fit.R
\name{plot.missSBM_fit}
\alias{plot.missSBM_fit}
\title{Visualization for an object \code{\link{missSBM_fit}}}
\usage{
\method{plot}{missSBM_fit}(x, type = c("data", "expected", "meso", "imputed", "monitoring"), ...)
}
\arguments{
\item{x}{an object with class \code{\link{missSBM_fit}}}

\item{type}{the type specifies the field to plot, either "data", "expected", "meso", "imputed" or "monitoring"}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
a ggplot object
}
\description{
Plot function for the various fields of a \code{\link{missSBM_fit}}: the fitted
SBM (network or connectivity), and a plot monitoring the optimization.
}
