% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM_fit-Class.R
\name{predicted.missSBM_fit}
\alias{predicted.missSBM_fit}
\alias{predict.missSBM_fit}
\title{Prediction of a ['missSBM_fit'] (i.e. network with imputed missing dyads)}
\usage{
\method{predict}{missSBM_fit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class [`missSBM_fit`]}

\item{...}{additional parameters for S3 compatibility.}
}
\value{
an adjacency matrix between pairs of nodes. Missing dyads are imputed with
their expected values, i.e. by there estimated probabilities of connection under the missing SBM.
}
\description{
Prediction of a ['missSBM_fit'] (i.e. network with imputed missing dyads)
}
