% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare network data for estimation with missing data}
\usage{
prepare_data(adjacencyMatrix, covariates = NULL,
  similarity = missSBM:::l1_similarity)
}
\arguments{
\item{adjacencyMatrix}{The adjacency matrix of the network (NAs allowed)}

\item{covariates}{An optional list with M entries (the M covariates). If the covariates are node-centred, each entry of \code{covariates}
must be a size-N vector;  if the covariates are dyad-centred, each entry of \code{covariates} must be N x N matrix.}

\item{similarity}{An optional R x R -> R function to compute similarities between node covariates. Default is \code{l1_similarity}, that is, -abs(x-y).
Only relevent when the \code{covariates} is a list of size-N vectors.}
}
\value{
Returns an R6 object with class \code{\link{sampledNetwork}}.
}
\description{
This function put together the adjacency matrix of a network and an optional list of covariates
into a single \code{\link{sampledNetwork}} object, ready to use for inference with the \code{\link{estimate}}
function of the missSBM package.
}
\examples{
data(war_graphs)
adj_beligerent <- war_graphs$beligerent \%>\% igraph::as_adj(sparse = FALSE)
sampledNet_war_nocov <- prepare_data(adj_beligerent)
military_power <- igraph::get.vertex.attribute(war_graphs$beligerent)$military_power
sampledNet_war_withcov <- prepare_data(adj_beligerent, list(military_power = military_power))
}
\seealso{
\code{\link{estimate}} and \code{\link{sampledNetwork}}.
}
