% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledNetwork-R6Class.R
\docType{data}
\name{sampledNetwork}
\alias{sampledNetwork}
\title{a sampled network}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sampledNetwork
}
\description{
a sampled network
}
\section{Fields}{

\describe{
\item{\code{samplingRate}}{}
}}

\keyword{datasets}
