% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-R6Class.R
\docType{data}
\name{sampling}
\alias{sampling}
\title{A class of function which modelisate the sampling design and then missing data in the graph.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sampling
}
\description{
A class of function which modelisate the sampling design and then missing data in the graph.
}
\section{Fields}{

\describe{
\item{\code{nNodes}}{number of nodes}

\item{\code{missingParam}}{sampling parameters}

\item{\code{blockProportion}}{vector of block proportion (a.k.a. alpha)}

\item{\code{modelParameters}}{vector of model parameters (a.k.a. theta)}
}}

\keyword{datasets}
