% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingSBM.R
\name{samplingSBM}
\alias{samplingSBM}
\title{Sampling of a network}
\usage{
samplingSBM(adjacencyMatrix, sampling, parameters, clusters = NULL)
}
\arguments{
\item{adjacencyMatrix}{The adjacency matrix of the network}

\item{sampling}{The sampling design used to sample the adjacency matrix}

\item{parameters}{The sampling parameters adapted to each sampling}

\item{clusters}{Clusters membership vector of the nodes, only necessary for class sampling, by default equal to NULL}
}
\value{
\code{samplingSBM} returns a matrix (the sampled adjacency matrix of the network given in parameter)
}
\description{
\code{samplingSBM} is a function that sample a matrix (the adjacency matrix of a network) under the SBM
}
\details{
The differents sampling designs are splitted into two families in which we find dyad-centered and node-centered samplings, for
more details see (\cite{1}) :\itemize{\item Missing At Random (MAR) \itemize{\item{edge: parameter = p
\deqn{p = P(Dyad (i,j) is sampled)}}
\item{node: parameter = p and
\deqn{p = P(Node i is sampled)}}
\item{snowball (one step):
like the MARNode sampling plus we sample neighbours of nodes sampled at the first batch}
}
\item Not Missing At Random (NMAR) \itemize{ \item{double_standard: parameter = (p0,p1) and
\deqn{p0 = P(Dyad (i,j) is sampled | the dyad is equal to 0)=}, p1 = P(Dyad (i,j) is sampled | the dyad is equal to 1)}
\item{degree: parameter = c(a,b) and
\deqn{logit(a+b*Degree(i)) = P(Node i is sampled | Degree(i))}}
\item{block: parameter = c(p(1),...,p(Q)) and
\deqn{p(q) = P(Node i is sampled | node i is in cluster q)}}
}}
}
\examples{
### A SBM model : ###
n <- 300
Q <- 3
alpha <- rep(1,Q)/Q                                                                # mixture parameter
pi <- diag(.45,Q) + .05                                                            # connectivity matrix
family <- "Bernoulli"                                                              # the emmission law
directed <- FALSE                                                                  # if the network is directed or not
mySBM <- simulateSBM(n, alpha, pi, family, directed)                               # simulation of ad Bernoulli non-directed SBM

### Results : ###
adjacencyMatrix <- mySBM$adjacencyMatrix                                           # the adjacency matrix

## Sampling of the data : ##
samplingParameters <- .5                                                           # the sampling rate
sampling <- "edge"                                                              # the sampling design
sampledNetwork <- samplingSBM(adjacencyMatrix, sampling, samplingParameters)     # the sampled adjacency matrix


}
\references{
[1] Tabouy, P. Barbillon, J. Chiquet. Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.
}
\seealso{
\code{\link{inferSBM}} and \code{\link{samplingSBM}}
}
\author{
T. Tabouy
}
