% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM-R6Class.R
\docType{data}
\name{SBM}
\alias{SBM}
\title{A SBM model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SBM
}
\arguments{
\item{nNodes}{The number of nodes}

\item{mixtureParam}{The mixture weights (commonly named alpha)}

\item{connectParam}{The connectivity matrix, i.e. inter and intra clusters parameters}
}
\description{
A R6-class that encode the model of SBM that have generated the data
}
\keyword{datasets}
