% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM-R6Class.R
\docType{data}
\name{SBM_BernoulliUndirected}
\alias{SBM_BernoulliUndirected}
\title{blabla}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SBM_BernoulliUndirected
}
\description{
blabla
}
\section{Fields}{

\describe{
\item{\code{completeLogLik}}{a function to compute the completed log-likelihood}
}}

\keyword{datasets}
