% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_collection-R6Class.R
\docType{data}
\name{SBM_collection}
\alias{SBM_collection}
\title{A collection of adjusted Stochastic Block Model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SBM_collection
}
\arguments{
\item{sampledNetwork}{The sampled network data (a square matrix)}

\item{vBlocks}{The vector of number of blocks considered in the collection}

\item{sampling}{The sampling design for missing data modeling : MAREdge, doubleStandard, MARNode, snowball, starDegree, class}

\item{family}{The emission law of the adjacency matrix : Bernoulli or Poisson}

\item{directed}{Boolean variable to indicate whether the network is directed or not,
by default "undirected" is choosen}
}
\description{
A R6-class that creates a collection of SBM with different numbers of groups in order to choose the number of
groups the fit at best the model (ICL criterion)
}
\keyword{datasets}
