% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_fit-Class.R
\name{SBM_fit}
\alias{SBM_fit}
\title{R6 Class definition of an SBM-fit}
\description{
R6 Class definition of an SBM-fit

R6 Class definition of an SBM-fit
}
\details{
This class is designed to adjust a Stochastic Block Model on a fully observed network.
The doVEM method performs inference via Variational EM.

This class is virtual: inference is effective only for instance of one of the two child
classes 'SBM_fit_nocovariate' and 'SBM_fit_covariates'
}
\section{Super class}{
\code{\link[missSBM:SBM]{missSBM::SBM}} -> \code{SBM_fit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{loglik}}{double: approximation of the log-likelihood (variational lower bound) reached}

\item{\code{ICL}}{double: value of the integrated classification log-likelihood}

\item{\code{probMemberships}}{matrix for clustering memberships}

\item{\code{memberships}}{vector of clustering}

\item{\code{penalty}}{double, value of the penalty term in ICL}

\item{\code{entropy}}{double, value of the entropy due to the clustering distribution}

\item{\code{expectation}}{expected values of connection under the current model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-doVEM}{\code{SBM_fit$doVEM()}}
\item \href{#method-show}{\code{SBM_fit$show()}}
\item \href{#method-clone}{\code{SBM_fit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="SBM" data-id="initialize">}\href{../../missSBM/html/SBM.html#method-initialize}{\code{missSBM::SBM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="SBM" data-id="plot">}\href{../../missSBM/html/SBM.html#method-plot}{\code{missSBM::SBM$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="SBM" data-id="print">}\href{../../missSBM/html/SBM.html#method-print}{\code{missSBM::SBM$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doVEM"></a>}}
\if{latex}{\out{\hypertarget{method-doVEM}{}}}
\subsection{Method \code{doVEM()}}{
method to perform estimation via variational EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_fit$doVEM(threshold = 1e-04, maxIter = 10, fixPointIter = 3, trace = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{stop when an optimization step changes the objective function by less than threshold. Default is 1e-4.}

\item{\code{maxIter}}{V-EM algorithm stops when the number of iteration exceeds maxIter. Default is 10}

\item{\code{fixPointIter}}{number of fix-point iterations in the Variational E step. Default is 3.}

\item{\code{trace}}{logical for verbosity. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
show method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_fit$show(type = "Fit of a Stochastic Block Model\\n")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character used to specify the type of SBM}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
