% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_VEMfit-R6Class.R
\docType{data}
\name{SBM_VEMfit}
\alias{SBM_VEMfit}
\title{an SBM fit, i.e. an adjusted SBM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SBM_VEMfit
}
\description{
an SBM fit, i.e. an adjusted SBM
}
\section{Fields}{

\describe{
\item{\code{ICL}}{}
}}

\keyword{datasets}
