% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSBM_fit_missSBM-Class.R
\name{SimpleSBM_fit_missSBM}
\alias{SimpleSBM_fit_missSBM}
\title{This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.}
\description{
This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.

This internal class is designed to adjust a binary Stochastic Block Model in the context of missSBM.
}
\details{
It is not designed not be call by the user
}
\section{Super classes}{
\code{\link[sbm:SBM]{sbm::SBM}} -> \code{\link[sbm:SBM_fit]{sbm::SBM_fit}} -> \code{\link[sbm:SimpleSBM_fit]{sbm::SimpleSBM_fit}} -> \code{SimpleSBM_fit_missSBM}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{vExpec}}{double: variational approximation of the expectation complete log-likelihood}

\item{\code{loglik}}{double: approximation of the log-likelihood (variational lower bound) reached}

\item{\code{ICL}}{double: value of the integrated classification log-likelihood}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimpleSBM_fit_missSBM$new()}}
\item \href{#method-doVEM}{\code{SimpleSBM_fit_missSBM$doVEM()}}
\item \href{#method-update_parameters}{\code{SimpleSBM_fit_missSBM$update_parameters()}}
\item \href{#method-update_blocks}{\code{SimpleSBM_fit_missSBM$update_blocks()}}
\item \href{#method-clone}{\code{SimpleSBM_fit_missSBM$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SBM" data-id="plot">}\href{../../sbm/html/SBM.html#method-plot}{\code{sbm::SBM$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SBM" data-id="print">}\href{../../sbm/html/SBM.html#method-print}{\code{sbm::SBM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SBM_fit" data-id="setModel">}\href{../../sbm/html/SBM_fit.html#method-setModel}{\code{sbm::SBM_fit$setModel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SimpleSBM_fit" data-id="optimize">}\href{../../sbm/html/SimpleSBM_fit.html#method-optimize}{\code{sbm::SimpleSBM_fit$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SimpleSBM_fit" data-id="predict">}\href{../../sbm/html/SimpleSBM_fit.html#method-predict}{\code{sbm::SimpleSBM_fit$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SimpleSBM_fit" data-id="reorder">}\href{../../sbm/html/SimpleSBM_fit.html#method-reorder}{\code{sbm::SimpleSBM_fit$reorder()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sbm" data-topic="SimpleSBM_fit" data-id="show">}\href{../../sbm/html/SimpleSBM_fit.html#method-show}{\code{sbm::SimpleSBM_fit$show()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for simpleSBM_fit for missSBM purpose
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimpleSBM_fit_missSBM$new(adjacencyMatrix, clusterInit, covarList = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adjacencyMatrix}}{a matrix encoding the graph}

\item{\code{clusterInit}}{Initial clustering: either a character in "hierarchical", "spectral" or "kmeans", or a vector with size \code{ncol(adjacencyMatrix)}, providing a user-defined clustering with \code{nbBlocks} levels. Default is "hierarchical".}

\item{\code{covarList}}{An option list with M entries (the M covariates).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doVEM"></a>}}
\if{latex}{\out{\hypertarget{method-doVEM}{}}}
\subsection{Method \code{doVEM()}}{
method to perform estimation via variational EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimpleSBM_fit_missSBM$doVEM(
  threshold = 1e-04,
  maxIter = 10,
  fixPointIter = 3,
  trace = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{stop when an optimization step changes the objective function by less than threshold. Default is 1e-4.}

\item{\code{maxIter}}{V-EM algorithm stops when the number of iteration exceeds maxIter. Default is 10}

\item{\code{fixPointIter}}{number of fix-point iterations in the Variational E step. Default is 3.}

\item{\code{trace}}{logical for verbosity. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-update_parameters}{}}}
\subsection{Method \code{update_parameters()}}{
update parameters estimation (M-step)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimpleSBM_fit_missSBM$update_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_blocks"></a>}}
\if{latex}{\out{\hypertarget{method-update_blocks}{}}}
\subsection{Method \code{update_blocks()}}{
update variational estimation of blocks (VE-step)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimpleSBM_fit_missSBM$update_blocks(log_lambda = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_lambda}}{double use to adjust the parameter estimation according to the sampling design}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimpleSBM_fit_missSBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
