% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulation of an SBM}
\usage{
simulate(nNodes, mixtureParam, connectParam, directed = FALSE,
  covariates = NULL, covarParam = NULL, similarity = l1_similarity)
}
\arguments{
\item{nNodes}{The number of nodes}

\item{mixtureParam}{The mixture parameters}

\item{connectParam}{The connectivity matrix (inter/intra clusters probabilities. provided on a logit scale for a model with covariates)}

\item{directed}{Boolean variable to indicate whether the network is directed or not. Default to \code{FALSE}.}

\item{covariates}{A list with M entries (the M covariates). If the covariates are node-centred, each entry of \code{covariates}
must be a size-N vector;  if the covariates are dyad-centred, each entry of \code{covariates} must be N x N matrix.}

\item{covarParam}{An optional vector of parameters associated with the covariates, with size M}

\item{similarity}{An optional R x R -> R function to compute similarities between node covariates. Default is \code{l1_similarity}, that is, -abs(x-y). Only relevent when covariates is a list of size-N vectors.}
}
\value{
an object with class \code{SBM_sampler}
}
\description{
Generates a realization (blocks and adjancency matrix) of a Stochastic Block model
}
\examples{
## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
M <- 2 # two Gaussian covariates
alpha <- rep(1,Q)/Q     # mixture parameters
pi <- diag(.45,Q) + .05 # connectivity matrix
eta <- rnorm(M, -1, 1)  # covaraite parametes
gamma <- log(pi/(1-pi)) # logit transform of pi for the model with covariates
X   <- replicate(M, rnorm(N,mean = 0, sd = 1), simplify = FALSE)
phi <- replicate(M, matrix(rnorm(N * N ,mean = 0, sd = 1), N, N), simplify = FALSE)
## draw a SBM without covariates
sbm <- missSBM::simulate(N, alpha, pi, directed)

## draw a SBM model with node-centred covariates
sbm_cov1 <- missSBM::simulate(N, alpha, gamma, directed, X, eta)

## draw a SBM model with dyad-centered covariates
sbm_cov2 <- missSBM::simulate(N, alpha, gamma, directed, X, eta)

\dontrun{
par(mfrow = c(1,3))
plot(sbm)
plot(sbm_cov1)
plot(sbm_cov2)
}

}
\seealso{
The class \code{\link{SBM_sampler}}
}
