% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Spectral Clustering}
\usage{
spectral_clustering(A, vBlocks)
}
\arguments{
\item{A}{a sparse Matrix with class 'dgCMatrix'}

\item{vBlocks}{a vector of integer containing the successive
number of clusters considered}
}
\description{
Spectral Clustering
}
