% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spectral_clustering_cpp}
\alias{spectral_clustering_cpp}
\title{Absolute Spectral Clustering}
\usage{
spectral_clustering_cpp(A, vBlocks)
}
\arguments{
\item{A}{a dgCMatrix}

\item{vBlocks}{a vector of integer for the succesive number of blocks considered}
}
\value{
a list of vector of clustering memberships
}
\description{
Absolute Spectral Clustering
}
