% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM_fit-Class.R
\name{summary.missSBM_fit}
\alias{summary.missSBM_fit}
\title{Summary method for a ['missSBM_fit']}
\usage{
\method{summary}{missSBM_fit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class [`missSBM_fit`]}

\item{...}{additional parameters for S3 compatibility.}
}
\description{
Summary method for a ['missSBM_fit']
}
