% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war_graphs.R
\docType{data}
\name{war_graphs}
\alias{war_graphs}
\title{War_graphs data set}
\format{A list with 2 two igraph objects, \code{alliance} and \code{beligerent}.}
\source{
networs were extracted from <http://www.correlatesofwar.org/>
}
\usage{
war_graphs
}
\description{
This dataset contains two networks where the nodes are countries and an edge in network "beligerent"
means that the two countries have been at least once at war between years 1816 to 2007 while an edge in network "alliance"
means that the two countries have had a formal alliance between years 1816 to 2012. The network `beligenrent` have less nodes since countries which have not been at war are not considered.
}
\examples{
data(war_graphs)
class(war_graphs$beligerent)
igraph::gorder(war_graphs$alliance)
igraph::edges(war_graphs$alliance)
}
\references{
Sarkees, Meredith Reid and Frank Wayman (2010). Resort to War: 1816 - 2007. Washington DC: CQ Press.

Gibler, Douglas M. 2009. International military alliances, 1648-2008. CQ Press
}
\keyword{datasets}
