% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampler-Class.R
\name{doubleStandardSampler}
\alias{doubleStandardSampler}
\title{Class for defining a double-standard sampler}
\description{
Class for defining a double-standard sampler

Class for defining a double-standard sampler
}
\section{Super classes}{
\code{\link[missSBM:networkSampling]{missSBM::networkSampling}} -> \code{\link[missSBM:networkSampler]{missSBM::networkSampler}} -> \code{\link[missSBM:dyadSampler]{missSBM::dyadSampler}} -> \code{doubleStandardSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{doubleStandardSampler$new()}}
\item \href{#method-clone}{\code{doubleStandardSampler$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSampling" data-id="print">}\href{../../missSBM/html/networkSampling.html#method-print}{\code{missSBM::networkSampling$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSampling" data-id="show">}\href{../../missSBM/html/networkSampling.html#method-show}{\code{missSBM::networkSampling$show()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSampler" data-id="rSamplingMatrix">}\href{../../missSBM/html/networkSampler.html#method-rSamplingMatrix}{\code{missSBM::networkSampler$rSamplingMatrix()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for networkSampling
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{doubleStandardSampler$new(parameters = NA, adjMatrix = NA, directed = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{the vector of parameters associated to the sampling at play}

\item{\code{adjMatrix}}{matrix of adjacency}

\item{\code{directed}}{logical, directed network of not}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{doubleStandardSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
