% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampling_fit-Class.R
\name{doubleStandardSampling_fit}
\alias{doubleStandardSampling_fit}
\title{Class for fitting a double-standard sampling}
\description{
Class for fitting a double-standard sampling

Class for fitting a double-standard sampling
}
\section{Super classes}{
\code{\link[missSBM:networkSampling]{missSBM::networkSampling}} -> \code{\link[missSBM:networkSamplingDyads_fit]{missSBM::networkSamplingDyads_fit}} -> \code{doubleStandardSampling_fit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{vExpec}}{variational expectation of the sampling}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{doubleStandardSampling_fit$new()}}
\item \href{#method-update_parameters}{\code{doubleStandardSampling_fit$update_parameters()}}
\item \href{#method-update_imputation}{\code{doubleStandardSampling_fit$update_imputation()}}
\item \href{#method-clone}{\code{doubleStandardSampling_fit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSampling" data-id="print">}\href{../../missSBM/html/networkSampling.html#method-print}{\code{missSBM::networkSampling$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSamplingDyads_fit" data-id="show">}\href{../../missSBM/html/networkSamplingDyads_fit.html#method-show}{\code{missSBM::networkSamplingDyads_fit$show()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{doubleStandardSampling_fit$new(sampledNetwork, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampledNetwork}}{a object with class sampledNetwork representing the observed data with possibly missing entries}

\item{\code{...}}{used for compatibility}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-update_parameters}{}}}
\subsection{Method \code{update_parameters()}}{
a method to update the estimation of the parameters. By default, nothing to do (corresponds to MAR sampling)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{doubleStandardSampling_fit$update_parameters(imputedNet, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{imputedNet}}{an adjacency matrix where missing values have been imputed}

\item{\code{...}}{use for compatibility}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_imputation"></a>}}
\if{latex}{\out{\hypertarget{method-update_imputation}{}}}
\subsection{Method \code{update_imputation()}}{
a method to update the imputation of the missing entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{doubleStandardSampling_fit$update_imputation(PI)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PI}}{the matrix of inter/intra class probability of connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{doubleStandardSampling_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
