% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampling_fit-Class.R
\docType{data}
\name{dyadSampling_fit}
\alias{dyadSampling_fit}
\title{Definition of R6 Class 'networkSampling_fit'}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dyadSampling_fit
}
\description{
This class is use to define a fit for a networkSampling. Inherits from 'networkSampling'
}
\keyword{datasets}
