% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Inference of an SBM with missing data}
\usage{
estimate(adjacencyMatrix, vBlocks, sampling,
  clusterInit = ifelse(is.null(covarMatrix), "hierarchical", "spectral"),
  covarMatrix = NULL, covarSimilarity = l1_similarity, trace = TRUE,
  smoothing = c("none", "forward", "backward", "both"), mc.cores = 1,
  iter_both = 1, control_VEM = list())
}
\arguments{
\item{adjacencyMatrix}{The adjacency matrix of the network}

\item{vBlocks}{The vector of number of blocks considered in the collection}

\item{sampling}{The sampling design for missing data modeling : "dyad", "double_standard", "node", "snowball", "degree", "block" by default "undirected" is choosen}

\item{clusterInit}{Initial method for clustering: either a character in "hierarchical", "spectral" or "kmeans", or a list with \code{length(vBlocks)} vectors, each with size \code{ncol(adjacencyMatrix)} providing a user-defined clustering}

\item{covarMatrix}{An optional matrix of covariates with dimension N x M (M covariates per node).}

\item{covarSimilarity}{An optional R x R -> R function  to compute similarity between node covariates. Default is #'}

\item{trace}{logical, control the verbosity. Default to \code{TRUE}.}

\item{smoothing}{character indicating what kind of ICL smoothing should be use among "none", "forward", "backward" or "both"}

\item{mc.cores}{integer, the number of cores to use when multiply model are fitted}

\item{iter_both}{integer for the number of iteration in case of foward-backward (aka both) smoothing}

\item{control_VEM}{a list controlling the variational EM algorithm. See details.}

\item{Robject}{an object with class \code{missSBMcollection}}
}
\value{
Returns an R6 object with class \code{\link{missSBM_collection}}.
}
\description{
Perform variational inference of a Stochastic Block Model from a sampled adjacency matrix
}
\examples{
## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix

## simulate a SBM without covariates
sbm <- missSBM::simulate(N, alpha, pi, directed)

## Sample network data
samplingParameters <- .5 # the sampling rate
sampling <- "dyad"       # the sampling design
sampledNet <- missSBM::sample(sbm$adjMatrix, sampling, samplingParameters)

## Inference :
vBlocks <- 1:5 # number of classes
collection <- missSBM::estimate(sampledNet$adjMatrix, vBlocks, sampling)
collection$ICL
}
\seealso{
\code{\link{sample}}, \code{\link{simulate}}, \code{\link{missSBM_collection}} and \code{\link{missingSBM_fit}}.
}
