% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Inference of an SBM with missing data}
\usage{
estimate(sampledNet, vBlocks, sampling, clusterInit = "hierarchical",
  control = list())
}
\arguments{
\item{sampledNet}{An object with class \code{\link{sampledNetwork}}, typically obtained wthe the function \code{\link{prepare_data}} or \code{\link{sample}}.}

\item{vBlocks}{The vector of number of blocks considered in the collection}

\item{sampling}{The sampling design for missing data modeling : "dyad", "node", "double-standard", "block-dyad", "block-node" ,"degree}

\item{clusterInit}{Initial method for clustering: either a character in "hierarchical", "spectral" or "kmeans", or a list with \code{length(vBlocks)} vectors, each with size \code{ncol(adjacencyMatrix)} providing a user-defined clustering}

\item{control}{a list controlling the variational EM algorithm. See details.}

\item{covariates}{A list with M entries (the M covariates). If the covariates are node-centred, each entry of \code{covariates}
must be a size-N vector;  if the covariates are dyad-centred, each entry of \code{covariates} must be N x N matrix.}

\item{similarity}{An optional R x R -> R function to compute similarities between node covariates. Default is \code{l1_similarity}, that is, -abs(x-y).
Only relevent when the covariates are node-centered (i.e. \code{covariates} is a list of size-N vectors).}
}
\value{
Returns an R6 object with class \code{\link{missSBM_collection}}.
}
\description{
Perform variational inference of a Stochastic Block Model from a sampled adjacency matrix
}
\details{
The list of parameters \code{control} controls the optimziation process and the variational EM algorithm, with the following entries
 \itemize{
 \item{"threshold"}{stop when an optimization step changes the objective function by less than threshold. Default is 1e-4.}
 \item{"maxiter"}{V-EM algorithm stops when the number of iteration exceeds maxIter. Default is 200}
 \item{"fixPointIter"}{number of fix-point iteration for the Variational E step. Default is 5.}
 \item{"cores"}{integer for number of cores used. Default is 1.}
 \item{"trace"}{integer for verbosity (0, 1, 2). Useless when \code{cores} > 1}
}
}
\examples{
## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix

## simulate a SBM without covariates
sbm <- missSBM::simulate(N, alpha, pi, directed)

## Sample network data
samplingParameters <- .5 # the sampling rate
sampling <- "dyad"       # the sampling design
sampledNet <- missSBM::sample(sbm$adjacencyMatrix, sampling, samplingParameters)

## Inference :
vBlocks <- 1:5 # number of classes
collection <- missSBM::estimate(sampledNet, vBlocks, sampling)
collection$ICL
}
\seealso{
\code{\link{sample}}, \code{\link{simulate}}, \code{\link{missSBM_collection}} and \code{\link{missSBM_fit}}.
}
