% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-missSBM_fit.R
\name{fitted.missSBM_fit}
\alias{fitted.missSBM_fit}
\title{Extract model fitted values from object  \link{'missSBM_fit'}, return by \code{\link[=estimateMissSBM]{estimateMissSBM()}}}
\usage{
\method{fitted}{missSBM_fit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{missSBM_fit}}}

\item{...}{additional parameters for S3 compatibility.}
}
\value{
A matrix of estimated probabilities of connection
}
\description{
Extract model fitted values from object  \link{'missSBM_fit'}, return by \code{\link[=estimateMissSBM]{estimateMissSBM()}}
}
