% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingSBM.R
\name{inferSBM}
\alias{inferSBM}
\title{Inference of Stochastic Block Model from sampled data}
\usage{
inferSBM(sampledNetwork, vBlocks, sampling, plot = TRUE)
}
\arguments{
\item{sampledNetwork}{The sampled network data (a square matrix)}

\item{vBlocks}{The vector of number of blocks considered in the collection}

\item{sampling}{The sampling design for missing data modeling : MAREdge, doubleStandard, MARNode, snowball, starDegree, class
by default "undirected" is choosen}

\item{plot}{Summary of the output of the algorithm, by default TRUE is choosen}
}
\value{
\code{inferSBM} returns a list with the best model choosen following the ICL criterion, a list with all models estimated for all Q in vBlocks
and a vector with ICL calculated for all Q in vBlocks
}
\description{
\code{inferSBM} is a function that makes variationnal inference of Stochastic Block Model from sampled adjacency matrix
}
\examples{
### A SBM model : ###
n <- 300
Q <- 3
alpha <- rep(1,Q)/Q                                                                # mixture parameter
pi <- diag(.45,Q) + .05                                                            # connectivity matrix
family <- "Bernoulli"                                                              # the emmission law
directed <- FALSE                                                                  # if the network is directed or not
mySBM <- simulateSBM(n, alpha, pi, family, directed)                               # simulation of ad Bernoulli non-directed SBM

### Results : ###
adjacencyMatrix <- mySBM$adjacencyMatrix                                           # the adjacency matrix

## Sampling of the data : ##
samplingParameters <- .5                                                           # the sampling rate
sampling <- "MAREdge"                                                              # the sampling design
sampledAdjMatrix <- samplingSBM(adjacencyMatrix, sampling, samplingParameters)     # the sampled adjacency matrix

## Inference :
vBlocks <- 1:5                                                                     # number of classes
sbm <- inferSBM(sampledAdjMatrix, vBlocks, sampling, family, directed)             # the inference

}
\references{
[1] Tabouy, P. Barbillon, J. Chiquet. Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.
}
\seealso{
\code{\link{samplingSBM}} and \code{\link{simulateSBM}} and \code{\link{SBM_collection}}.
}
\author{
T. Tabouy
}
