% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingSBM.R
\name{missingSBM}
\alias{missingSBM}
\title{Generates \code{\link{SBM_collection}} objects.}
\usage{
missingSBM(sampledNetwork, vBlocks, sampling, family, directed)
}
\arguments{
\item{sampledNetwork}{The sampled network data (a square matrix)}

\item{vBlocks}{The vector of number of blocks considered in the collection}

\item{sampling}{The sampling design for missing data modeling : MAREdge, doubleStandard, MARNode, snowball, starDegree, class}

\item{family}{The emission law of the adjacency matrix : Bernoulli or Poisson}

\item{directed}{Boolean variable to indicate whether the network is directed or not,
by default "undirected" is choosen}
}
\value{
\code{missingSBM} returns a \code{\link{SBM_collection}} object.
}
\description{
\code{missingSBM} is a function that makes variationnal inference of Stochastic Block Model from sampled adjacency matrix
}
\examples{

}
\references{
[1] Tabouy et al., Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.
}
\seealso{
\code{\link{SBM_collection}}.
}
\author{
T. Tabouy
}
