% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM-package.R
\docType{package}
\name{missSBM}
\alias{missSBM}
\alias{missSBM-package}
\title{An R package for adjusting Stochastic Block Models from networks data sampled under various missing data conditions}
\description{
The SBM package provides three categories of important functions:
\code{\link{simulateSBM}}, \code{\link{samplingSBM}}, \code{\link{inferSBM}}.
}
\references{
[1] Tabouy et al., Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.

[2] Kolaczyk Erik D. (2009). Statistical Analysis of Network Data: Methods and Models. Springer.
}
\author{
Timothée Tabouy, Pierre Barbillon, Julien Chiquet

Maintainer: \email{timothee.tabouy@gmail.com}
}
