% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM-package.R
\docType{package}
\name{missSBM}
\alias{missSBM}
\alias{missSBM-package}
\title{An R package for adjusting Stochastic Block Models from networks data sampled under various missing data conditions}
\description{
The SBM package provides three categories of important functions:
\code{\link{simulateSBM}}, \code{\link{samplingSBM}}, \code{\link{inferSBM}}.
}
\details{
Package: missSBM

Type:    Package

Version: 0.1.0

Date:    2017-12-01

License: GPL-2 | GPL-3
}
\references{
[1] Tabouy et al., Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.

[2] Kolaczyk Erik D. (2009). Statistical Analysis of Network Data: Methods and Models. Springer.
}
\author{
Timothée Tabouy

Maintainer: \email{timothee.tabouy@gmail.com}
}
