% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM-package.R
\docType{package}
\name{missSBM}
\alias{missSBM}
\alias{missSBM-package}
\title{Bayesian calibration for computational codes}
\description{
The foo package provides three categories of important functions:
\code{\link{nomADonner}}.
}
\details{
Package: missSBM

Type:    Package

Version: 0.1.0

Date:    2017-12-01

License: GPL-2 | GPL-3
}
\examples{
# Introduction to missSBM
vignette("missSBM-introduction")

# information on how to use the package
vignette("missSBM-package")
}
\references{
Kolaczyk Erik D. (2009). Statistical Analysis of Network Data: Methods and Models. Springer.
}
\author{
Timothée Tabouy

Maintainer: \email{timothee.tabouy@gmail.com}
}
