% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods_missSBMcollection.R,
%   R/missSBM-package.R, R/missSBM.R
\docType{package}
\name{is.missSBMcollection}
\alias{is.missSBMcollection}
\alias{ICL}
\alias{ICL.missSBMcollection}
\alias{optimizationStatus}
\alias{optimizationStatus.missSBMcollection}
\alias{getBestModel}
\alias{getBestModel.missSBMcollection}
\alias{missSBM}
\alias{missSBM-package}
\title{An R package for adjusting Stochastic Block Models from networks data sampled under various missing data conditions}
\usage{
is.missSBMcollection(Robject)

ICL(Robject)

\method{ICL}{missSBMcollection}(Robject)

optimizationStatus(Robject)

\method{optimizationStatus}{missSBMcollection}(Robject)

getBestModel(Robject)

\method{getBestModel}{missSBMcollection}(Robject)

missSBM(adjacencyMatrix, vBlocks, sampling,
  clusterInit = ifelse(is.null(covarMatrix), "hierarchical", "spectral"),
  covarMatrix = NULL, covarSimilarity = l1_similarity, trace = TRUE,
  smoothing = c("none", "forward", "backward", "both"), mc.cores = 1,
  iter_both = 1, control_VEM = list())
}
\arguments{
\item{Robject}{an object with class \code{missSBMcollection}}

\item{adjacencyMatrix}{The adjacency matrix of the network}

\item{vBlocks}{The vector of number of blocks considered in the collection}

\item{sampling}{The sampling design for missing data modeling : "dyad", "double_standard", "node", "snowball", "degree", "block"
by default "undirected" is choosen}

\item{covarMatrix}{An optional matrix of covariates with dimension N x M (M covariates per node).}

\item{covarSimilarity}{An optional R x R -> R function  to compute similarity between node covariates. Default is #' @param clusterInit character in "hierarchical" or "spectral" for initialization}

\item{trace}{logical, control the verbosity. Default to \code{TRUE}.}

\item{smoothing}{character indicating what kind of ICL smoothing should be use among "none", "forward", "backward" or "both"}

\item{mc.cores}{integer, the number of cores to use when multiply model are fitted}

\item{iter_both}{integer for the number of iteration in case of foward-backward (aka both) smoothing}

\item{control_VEM}{a list controlling the variational EM algorithm. See details.}
}
\value{
Returns an S3 object with class \code{missSBMcollection}, which is a list with all models estimated for all Q in vBlocks. \code{missSBMcollection} owns a couple of S3 methods: \code{is.missSBMcollection} to test the class of the object, a method \code{ICL} to extract the values of the Integrated Classification Criteria for each model, a method \code{getBestModel} which extract from the list the best model (and object of class \code{missSBM-fit}) according to the ICL, and a method \code{optimizationStatus} to monitor the objective function a convergence of the VEM algorithm.
}
\description{
The SBM package provides three categories of important functions:
\code{\link{simulateSBM}}, \code{\link{sampleNetwork}}, \code{\link{missSBM}}.

Perform variational inference of Stochastic Block Model from sampled adjacency matrix
}
\examples{
### A SBM model : ###
N <- 300
Q <- 3
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix
directed <- FALSE
## Simulation of an Bernoulli non-directed SBM
mySBM <- simulateSBM(N, alpha, pi, directed)

## Sampling of the data : ##
samplingParameters <- .5 # the sampling rate
sampling <- "dyad"       # the sampling design
sampledNet <-
   sampleNetwork(
     mySBM$adjMatrix,
     sampling,
     samplingParameters
   )

## Inference :
vBlocks <- 1:5 # number of classes
sbm <-
   missSBM(
      sampledNet$adjMatrix,
      vBlocks,
      sampling
   )
}
\references{
[1] Tabouy et al., Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.

[2] Kolaczyk Erik D. (2009). Statistical Analysis of Network Data: Methods and Models. Springer.

[1] Tabouy, P. Barbillon, J. Chiquet. Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.
}
\seealso{
\code{\link{sampleNetwork}} and \code{\link{simulateSBM}} and \code{\link{missingSBM_fit}}.
}
\author{
Timothée Tabouy, Pierre Barbillon, Julien Chiquet
}
