% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM_collection-Class.R
\name{missSBM_collection}
\alias{missSBM_collection}
\title{An object to represent a collection of missSBM_fit}
\description{
This R6 class stores a collection of missSBM_fit. Comes with basic printing methods an
field access.
}
\seealso{
The function \code{\link{estimate}}, which produces an instance of this class.
The function \code{\link{smooth}} can be used to smooth the ICL on a collection of model,
as post-treatment.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{models}}{a list of models}

\item{\code{ICL}}{the vector of Integrated Classification Criterion (ICL) associated to
the models in the collection (the smaller, the better)}

\item{\code{bestModel}}{the best model according to the ICL}

\item{\code{optimizationStatus}}{a data.frame summarizing the optimization process for all models}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{missSBM_collection$new()}}
\item \href{#method-estimate}{\code{missSBM_collection$estimate()}}
\item \href{#method-smooth}{\code{missSBM_collection$smooth()}}
\item \href{#method-show}{\code{missSBM_collection$show()}}
\item \href{#method-print}{\code{missSBM_collection$print()}}
\item \href{#method-clone}{\code{missSBM_collection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_collection$new(
  sampledNet,
  vBlocks,
  sampling,
  clusterInit,
  cores,
  trace,
  useCov
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-estimate}{}}}
\subsection{Method \code{estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_collection$estimate(control)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smooth"></a>}}
\if{latex}{\out{\hypertarget{method-smooth}{}}}
\subsection{Method \code{smooth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_collection$smooth(type, control)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_collection$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_collection$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_collection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
