% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM_collection-Class.R
\docType{data}
\name{missSBM_collection}
\alias{missSBM_collection}
\title{An object to represent a collection of missSBM_fit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
missSBM_collection
}
\description{
This R6 class stores a collection of missSBM_fit. Comes with basic printing methods an
field access.
}
\section{Fields}{

\describe{
\item{\code{models}}{a list of models}

\item{\code{ICL}}{the vector of Integrated Classfication Criterion (ICL) associated to
the models in the collection (the smaller, the better)}

\item{\code{bestModel}}{the best model according to the ICL}

\item{\code{optimizationStatus}}{a data.frame summarizing the optimization process for all models}
}}

\seealso{
The function \code{\link{estimate}}, which produces an instance of this class.
The function \code{\link{smooth}} can be used to smooth the ICL on a collection of model,
as post-treatment.
}
\keyword{datasets}
