% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBMcollection-Class.R
\docType{data}
\name{missSBM_collection}
\alias{missSBM_collection}
\title{R6 Class definition of an missSBMcollection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
missSBM_collection
}
\description{
This class is to store a collection fo missSBM-fit
}
\keyword{datasets}
