% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM_fit-Class.R
\name{missSBM_fit}
\alias{missSBM_fit}
\title{An R6 class to represent an SBM fit with missing data}
\description{
The function [estimate()] fits a collection of SBM for varying number of block.
Each fitted SBM is an instance of an R6 object with class ['missSBM_fit'], described here.

Fields are accessed via active binding and cannot be changed by the user.

This class comes with a set of R6 methods, some of them being useful for the user and exported
as S3 methods. See the documentation for  [show()], [print()], [fitted()], [predict()], [plot()].
}
\examples{
## Sample 75\% of dyads in  French political Blogosphere's network data
adjMatrix <- missSBM::frenchblog2007 \%>\%
  igraph::as_adj (sparse = FALSE) \%>\%
  missSBM::sample(sampling = "dyad", parameters = 0.25)
collection <- missSBM::estimate(adjMatrix, 3:5, sampling = "dyad")
my_missSBM_fit <- collection$bestModel
class(my_missSBM_fit)
plot(my_missSBM_fit, "connectivity")

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fittedSBM}}{the fitted SBM, inheriting from class [`SBM_fit`]}

\item{\code{fittedSampling}}{the fitted sampling, inheriting from class [`networkSampling_fit`]}

\item{\code{sampledNetwork}}{The original network data used for the fit, with class [`sampledNetwork`]}

\item{\code{imputedNetwork}}{The network data with NAs values imputed with the model.}

\item{\code{monitoring}}{a list carrying information about the optimization process}

\item{\code{entropyImputed}}{the entropy of the distribution of the imputed dyads}

\item{\code{loglik}}{double: approximation of the log-likelihood (variational lower bound) reached}

\item{\code{vExpec}}{double: variational expectation of the complete log-likelihood}

\item{\code{penalty}}{double, value of the penalty term in ICL}

\item{\code{ICL}}{double: value of the integrated classification log-likelihood}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{missSBM_fit$new()}}
\item \href{#method-doVEM}{\code{missSBM_fit$doVEM()}}
\item \href{#method-show}{\code{missSBM_fit$show()}}
\item \href{#method-print}{\code{missSBM_fit$print()}}
\item \href{#method-clone}{\code{missSBM_fit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor for networkSampling
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_fit$new(sampledNet, nbBlocks, netSampling, clusterInit, useCov)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampledNet}}{An object with class [`sampledNetwork`], typically obtained with the function [`prepare_data`] (real-word data) or [`sample`] (simulation).}

\item{\code{nbBlocks}}{integer, the number of blocks in the SBM}

\item{\code{netSampling}}{The sampling design for the modelling of missing data: MAR designs ("dyad", "node") and NMAR designs ("double-standard", "block-dyad", "block-node" ,"degree")}

\item{\code{clusterInit}}{Initial clustering: either a character in "hierarchical", "spectral" or "kmeans", or a vector with size \code{ncol(adjacencyMatrix)}, providing a user-defined clustering with \code{nbBlocks} levels. Default is "hierarchical".}

\item{\code{useCov}}{logical. If covariates are present in sampledNet, should they be used for the inference or of the network sampling design, or just for the SBM inference? default is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doVEM"></a>}}
\if{latex}{\out{\hypertarget{method-doVEM}{}}}
\subsection{Method \code{doVEM()}}{
a method to perform inference of the current missSBM fit with variational EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_fit$doVEM(
  control = list(threshold = 0.001, maxIter = 100, fixPointIter = 5, trace = 1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control}}{a list of parameters controlling the variational EM algorithm. See details of function [`estimate`]}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
show method for missSBM_fit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_fit$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
User friendly print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_fit$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{missSBM_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
