% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_ER}
\alias{net_ER}
\title{ESR ego centred network}
\format{A sparse symmetric matrix with 741 rows and 741 columns:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{https://string-db.org/}
}
\usage{
net_ER
}
\description{
A dataset containing the weigthed PPI network centred around ESR1 and extracted from string data base
diamonds.
}
\keyword{datasets}
