% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampling_fit-Class.R
\docType{data}
\name{networkSamplingDyads_fit}
\alias{networkSamplingDyads_fit}
\title{Virtual class used to define a family of networkSamplingDyads_fit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
networkSamplingDyads_fit
}
\description{
Virtual class used to define a family of networkSamplingDyads_fit
}
\keyword{datasets}
