% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampling_fit-Class.R
\docType{data}
\name{networkSamplingNodes_fit}
\alias{networkSamplingNodes_fit}
\title{Virtual class used to define a family of networkSamplingNodes_fit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
networkSamplingNodes_fit
}
\description{
Virtual class used to define a family of networkSamplingNodes_fit
}
\keyword{datasets}
