% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSampling_fit-Class.R
\name{nodeSampling_fit}
\alias{nodeSampling_fit}
\title{Class for fitting a node sampling}
\description{
Class for fitting a node sampling

Class for fitting a node sampling
}
\section{Super classes}{
\code{\link[missSBM:networkSampling]{missSBM::networkSampling}} -> \code{\link[missSBM:networkSamplingNodes_fit]{missSBM::networkSamplingNodes_fit}} -> \code{nodeSampling_fit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{vExpec}}{variational expectation of the sampling}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{nodeSampling_fit$new()}}
\item \href{#method-clone}{\code{nodeSampling_fit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSamplingNodes_fit" data-id="show">}\href{../../missSBM/html/networkSamplingNodes_fit.html#method-show}{\code{missSBM::networkSamplingNodes_fit$show()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSamplingNodes_fit" data-id="update_imputation">}\href{../../missSBM/html/networkSamplingNodes_fit.html#method-update_imputation}{\code{missSBM::networkSamplingNodes_fit$update_imputation()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="networkSamplingNodes_fit" data-id="update_parameters">}\href{../../missSBM/html/networkSamplingNodes_fit.html#method-update_parameters}{\code{missSBM::networkSamplingNodes_fit$update_parameters()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeSampling_fit$new(sampledNetwork, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampledNetwork}}{a object with class sampledNetwork representing the observed data with possibly missing entries}

\item{\code{...}}{used for compatibility}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nodeSampling_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
