% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBM_fit-Class.R
\name{plot.missSBM_fit}
\alias{plot.missSBM_fit}
\title{Visualization for an object [`missSBM_fit`]}
\usage{
\method{plot}{missSBM_fit}(
  x,
  type = c("network", "connectivity", "sampledNetwork", "monitoring"),
  ...
)
}
\arguments{
\item{x}{an object with class [`missSBM_fit`]}

\item{type}{the type specifies the field to plot, either "network", "connectivity", "sampledNetwork" of "monitoring"}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\description{
Plot function for the various fields of a [`missSBM_fit`]
}
