% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledNetwork-Class.R
\name{sampledNetwork}
\alias{sampledNetwork}
\title{An R6 Class to represent sampled network data}
\description{
The function \code{\link{sample}} and \code{\link{prepare_data}} produces an instance of an object with class \code{sampledNetwork}.
}
\details{
All fields of this class are only accessible for reading. This class comes with a basic plot, summary and print methods
}
\examples{
## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix

## simulate a SBM without covariates
sbm <- missSBM::simulate(N, alpha, pi, directed)

## Sample network data
sampled_network <-
     missSBM::sample(
       adjacencyMatrix = sbm$adjacencyMatrix,
       sampling        = "double-standard",
       parameters      = c(0.4, 0.8)
     )

summary(sampled_network)
print(sampled_network)
plot(sampled_network, clustering = sbm$memberships)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{samplingRate}}{percentage of observed dyads}

\item{\code{nNodes}}{number of nodes}

\item{\code{nDyads}}{number of dyads}

\item{\code{is_directed}}{direction}

\item{\code{adjacencyMatrix}}{adjacency matrix (with NA)}

\item{\code{covarMatrix}}{the matrix of covariates (if applicable)}

\item{\code{covarArray}}{the array of covariates (if applicable)}

\item{\code{dyads}}{list of potential dyads in the network}

\item{\code{missingDyads}}{array indices of missing dyads}

\item{\code{observedDyads}}{array indices of observed dyads}

\item{\code{samplingMatrix}}{matrix of observed and non-observed edges}

\item{\code{observedNodes}}{vector of observed and non-observed nodes}

\item{\code{NAs}}{boolean for NA entries in the adjacencyMatrix}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sampledNetwork$new()}}
\item \href{#method-plot}{\code{sampledNetwork$plot()}}
\item \href{#method-show}{\code{sampledNetwork$show()}}
\item \href{#method-print}{\code{sampledNetwork$print()}}
\item \href{#method-clone}{\code{sampledNetwork$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sampledNetwork$new(adjacencyMatrix, covarMatrix = NULL, covarArray = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sampledNetwork$plot(
  clustering = NULL,
  main = paste("Network with sampling rate:", signif(self$samplingRate, 3))
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sampledNetwork$show(model = "Sampled Network\\n")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sampledNetwork$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sampledNetwork$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
