% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledNetwork-R6Class.R
\docType{data}
\name{sampledNetwork}
\alias{sampledNetwork}
\title{A sampled network}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sampledNetwork
}
\arguments{
\item{adjacencyMatrix}{The sampled adjacency matrix}

\item{directed}{Boolean variable to indicate whether the network is directed or not}
}
\description{
A R6-class that create a sampled network object, it collects sampling informations from the sampled matrix and
 is used during the inference
}
\keyword{datasets}
