% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-R6Class.R
\docType{data}
\name{sampling}
\alias{sampling}
\title{an SBM model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sampling
}
\description{
an SBM model
}
\section{Fields}{

\describe{
\item{\code{nNodes}}{number of nodes}

\item{\code{nBlocks}}{number of blocks}

\item{\code{blockProportion}}{vector of block proportion (a.k.a. alpha)}

\item{\code{modelParameters}}{vector of model parameters (a.k.a. theta)}
}}

\keyword{datasets}
