% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-R6Class.R
\docType{data}
\name{sampling}
\alias{sampling}
\title{A sampling}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sampling
}
\arguments{
\item{nNodes}{Number of nodes}

\item{missingParam}{Sampling parameters}

\item{directed}{Boolean variable to indicate whether the network is directed or not}
}
\description{
A R6-class which modelisate the sampling design and then missing data in the graph
}
\keyword{datasets}
