% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM-Class.R
\name{SBM}
\alias{SBM}
\title{An R6 Class to represent a SBM}
\description{
An R6 Class to represent a SBM

An R6 Class to represent a SBM
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nNodes}}{The number of nodes}

\item{\code{nBlocks}}{The number of blocks}

\item{\code{nCovariates}}{The number of covariates}

\item{\code{nDyads}}{The number of dyads}

\item{\code{direction}}{character indicating if the network is directed or not}

\item{\code{hasCovariates}}{boolean indicating if the SBM owns covariate(s)}

\item{\code{mixtureParam}}{the vector of mixture parameters (block proportions)}

\item{\code{connectParam}}{the matrix of connectivity: inter/intra probabilities of connection when the network does not have covariates, or a logit scaled version of it.}

\item{\code{adjacencyMatrix}}{The adjacency matrix of the network}

\item{\code{covarParam}}{the vector of parameters associated with the covariates}

\item{\code{covarArray}}{the array of covariates}

\item{\code{df_mixtureParams}}{degrees of freedoms for the mixture parameters}

\item{\code{df_connectParams}}{degrees of freedoms for the connectivity parameters}

\item{\code{df_covarParams}}{degrees of freedoms for the covariate parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SBM$new()}}
\item \href{#method-show}{\code{SBM$show()}}
\item \href{#method-print}{\code{SBM$print()}}
\item \href{#method-plot}{\code{SBM$plot()}}
\item \href{#method-clone}{\code{SBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an [`SBM`] object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$new(
  directed = FALSE,
  nNodes = NA,
  mixtureParam = NA,
  connectParam = NA,
  covarParam = NULL,
  covarArray = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{directed}}{logical indicating if the network is directed or not}

\item{\code{nNodes}}{The number of nodes}

\item{\code{mixtureParam}}{the vector of mixture parameters}

\item{\code{connectParam}}{the matrix of connectivity: inter/intra probabilities of connection when the network does not have covariates, or a logit scaled version of it.}

\item{\code{covarParam}}{the vector of parameters associated with the covariates}

\item{\code{covarArray}}{the array of covariates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
show method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$show(type = "Stochastic Block Model\\n")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character used to specify the type of SBM}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
User friendly print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
basic matrix plot method for SBM object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$plot(type = c("network", "connectivity"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character for the type of plot: either 'network' (true connection) or 'connectivity' (fitted connection). Default to 'network'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
