% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_collection-R6Class.R
\docType{data}
\name{SBM_collection}
\alias{SBM_collection}
\title{A collection of adjusted Stochastic Block Model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SBM_collection
}
\description{
A collection of adjusted Stochastic Block Model
}
\keyword{datasets}
