% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_fit-Class.R
\docType{data}
\name{SBM_fit}
\alias{SBM_fit}
\title{R6 Class definition of an SBM-fit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SBM_fit
}
\description{
This class is designed to adjust a Stochastic Block Model on a fully observed network.
The doVEM method performs inference via Variational EM
}
\keyword{datasets}
