% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_sampler-Class.R
\docType{data}
\name{SBM_sampler}
\alias{SBM_sampler}
\title{R6 Class definition of a sampler for a Stochastic Block Model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SBM_sampler
}
\description{
this class is used to sample from an SBM, thus has some additional fields and methods related to the blocks and the adjacency matrix (Z and Y)
}
\keyword{datasets}
