% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_sampler-Class.R
\name{SBM_sampler}
\alias{SBM_sampler}
\title{An R6 Class to represent a sampler for a SBM}
\description{
An R6 Class to represent a sampler for a SBM

An R6 Class to represent a sampler for a SBM
}
\details{
The function [`simulate`] produces an instance of an object with class [`SBM_sampler`].

All fields of this class are only accessible for reading. This class comes with a set of methods,
some of them being useful for the user (see examples)
\itemize{
\item{R6 methods:}{$rMemberships(), $rAdjacency()}
\item{S3 methods}{print(), plot()}
}

Fields are accessed via active binding and cannot be changed by the user.
}
\examples{
## SBM parameters
directed <- FALSE
N <- 300 # number of nodes
Q <- 3   # number of clusters
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix
gamma <- log(pi/(1-pi)) # logit transform fo the model with covariates

## draw a SBM without covariates through simulateSBM
sbm <- missSBM::simulate(N, alpha, pi, directed)

## equivalent construction from the SBM_sampler class itslef
sbm_s <- SBM_sampler$new(directed, N, alpha, pi)
sbm_s$rMemberships() # draw some blocks
sbm_s$rAdjacency() # draw some edges

coef(sbm_s, "block")
coef(sbm_s, "connectivity")
summary(sbm_s)
}
\seealso{
The function \code{\link{simulate}}.
}
\section{Super class}{
\code{\link[missSBM:SBM]{missSBM::SBM}} -> \code{SBM_sampler}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{indMemberships}}{matrix for clustering memberships}

\item{\code{memberships}}{vector of clustering}

\item{\code{expectation}}{expected values of connection under the current model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SBM_sampler$new()}}
\item \href{#method-rMemberships}{\code{SBM_sampler$rMemberships()}}
\item \href{#method-rAdjacency}{\code{SBM_sampler$rAdjacency()}}
\item \href{#method-show}{\code{SBM_sampler$show()}}
\item \href{#method-print}{\code{SBM_sampler$print()}}
\item \href{#method-clone}{\code{SBM_sampler$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="missSBM" data-topic="SBM" data-id="plot">}\href{../../missSBM/html/SBM.html#method-plot}{\code{missSBM::SBM$plot()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a [`SBM_sampler`] object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_sampler$new(
  directed = FALSE,
  nbNodes = NA,
  blockProp = NA,
  connectParam = NA,
  covarParam = numeric(0),
  covarList = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{directed}}{logical indicating if the network is directed or not}

\item{\code{nbNodes}}{The number of nodes}

\item{\code{blockProp}}{the vector of mixture parameters}

\item{\code{connectParam}}{the matrix of connectivity: inter/intra probabilities of connection when the network does not have covariates, or a logit scaled version of it.}

\item{\code{covarParam}}{the vector of parameters associated with the covariates}

\item{\code{covarList}}{A list with M entries (the M covariates). Each entry of the list must be an N x N matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rMemberships"></a>}}
\if{latex}{\out{\hypertarget{method-rMemberships}{}}}
\subsection{Method \code{rMemberships()}}{
a method to generate a vector of clusters indicators
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_sampler$rMemberships()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rAdjacency"></a>}}
\if{latex}{\out{\hypertarget{method-rAdjacency}{}}}
\subsection{Method \code{rAdjacency()}}{
a method to sample an adjacency matrix for the current SBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_sampler$rAdjacency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
show method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_sampler$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
User friendly print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_sampler$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBM_sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
