% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_VEMfit-R6Class.R
\docType{data}
\name{SBM_VEMfit}
\alias{SBM_VEMfit}
\title{An adjusted SBM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SBM_VEMfit
}
\arguments{
\item{SBM}{A R6-class object SBM (of class SBM) which encode the model}

\item{sampledNetwork}{A R6-class object sampled Network (of class sampledNetwork) which encode the data}

\item{sampling}{A R6-class object sampling (of class sampling) which encode the sampling design}

\item{init}{By default doing a spectral clustering ("spectralC"), can also be "CAH" a hierachical clustering more robust than spectral clustering
with sparse networks}

\item{blockInit}{By default equal to NULL, ad-hoc clustering to initialize the algorithm}

\item{controlVEM}{By default equal to 1e-5, stop criterion to say that the algorithm has converged}

\item{maxIterVEM}{By default equal to 100, stop criterion}
}
\description{
A R6-class which apply a Variational E-M to a given SBM with a given number of groups.
}
\keyword{datasets}
