% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingSBM.R
\name{simulateSBM}
\alias{simulateSBM}
\title{Simulation of a Stochastic Block Model}
\usage{
simulateSBM(N, alpha, pi, directed = FALSE, covarMatrix = NULL,
  covarParam = NULL, covarSimilarity = l1_similarity)
}
\arguments{
\item{N}{The number of nodes}

\item{alpha}{The mixture parameters}

\item{pi}{The connectivity matrix (probabilities inter and intra clusters)}

\item{directed}{Boolean variable to indicate whether the network is directed or not,
by default "undirected" is choosen}

\item{covarMatrix}{An optional matrix of covariates with dimension N x M (M covariates per node).}

\item{covarParam}{An optional vector of parameters associated with the covariates, with size M}

\item{covarSimilarity}{An optional R x R -> R function  to compute similarity between node covariates. Default is \code{l1_similarity}, that is, -abs(x-y).}
}
\value{
a object with class \code{SBM-Class}
}
\description{
Generates realizations (blocks and adjancency) of a Stochastic Block model
}
\examples{
### A SBM model : ###
N <- 300
Q <- 3
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix
directed <- FALSE
## Simulation of an Bernoulli non-directed SBM
mySBM <- simulateSBM(N, alpha, pi, directed)

}
