% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingSBM.R
\name{simulateSBM}
\alias{simulateSBM}
\title{Simulation of a Stochastic Block Model}
\usage{
simulateSBM(n, alpha, pi, directed = FALSE)
}
\arguments{
\item{n}{The number of nodes}

\item{alpha}{The mixture parameters}

\item{pi}{The connectivity matrix (probabilities inter and intra clusters)}

\item{directed}{Boolean variable to indicate whether the network is directed or not,
by default "undirected" is choosen}
}
\value{
\code{simulateSBM} returns a vector with clusters of nodes and a matrix (the adjacency matrix of the network)
}
\description{
\code{simulateSBM} is a function that generates a matrix (the adjacency matrix of a network) under the SBM
}
\details{
The emission law can be :\itemize{\item{Bernoulli:
\deqn{P(Y[i,j] = 1 | Zi = q, Zj = l) = p(Zi,Zj)}}
\item{Poisson:
\deqn{P(Y[i,j] = k | Zi = q, Zj = l) = (\lambda(Z_i,Z_j)^k/(k!)) * exp(-\lambda(Z_i,Z_j))}}
}
}
\examples{
### A SBM model : ###
n <- 300
Q <- 3
alpha <- rep(1,Q)/Q                                      # mixture parameter
pi <- diag(.45,Q) + .05                                  # connectivity matrix
directed <- FALSE
mySBM <- simulateSBM(n, alpha, pi, directed)             # simulation of ad Bernoulli non-directed SBM

### Results : ###
clusters <-  mySBM$clusters                              # clusters
adjacencyMatrix <- mySBM$adjacencyMatrix                 # the adjacency matrix

}
\references{
[1] Tabouy, P. Barbillon, J. Chiquet. Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.
}
\seealso{
\code{\link{inferSBM}} and \code{\link{samplingSBM}}
}
