% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingSBM.R
\name{simulateSBM}
\alias{simulateSBM}
\title{Simulation of a Stochastic Block Model}
\usage{
simulateSBM(N, alpha, pi, directed = FALSE, covariates = NULL,
  covarParam = NULL)
}
\arguments{
\item{N}{The number of nodes}

\item{alpha}{The mixture parameters}

\item{pi}{The connectivity matrix (probabilities inter and intra clusters)}

\item{directed}{Boolean variable to indicate whether the network is directed or not,
by default "undirected" is choosen}

\item{covariates}{An optional matrix of covariates with dimension N x M (M covariates per node).}

\item{covarParam}{An optional vector of parameters associated with the covariates, with size M}
}
\value{
\code{simulateSBM} returns a vector with clusters of nodes and a matrix (the adjacency matrix of the network)
}
\description{
\code{simulateSBM} is a function that generates a matrix (the adjacency matrix of a network) under the SBM
}
\details{
The emission law can be :\itemize{\item{Bernoulli:
\deqn{P(Y[i,j] = 1 | Zi = q, Zj = l) = p(Zi,Zj)}}
\item{Poisson:
\deqn{P(Y[i,j] = k | Zi = q, Zj = l) = (\lambda(Z_i,Z_j)^k/(k!)) * exp(-\lambda(Z_i,Z_j))}}
}
}
\examples{
### A SBM model : ###
N <- 300
Q <- 3
alpha <- rep(1,Q)/Q     # mixture parameter
pi <- diag(.45,Q) + .05 # connectivity matrix
directed <- FALSE
## Simulation of an Bernoulli non-directed SBM
mySBM <- simulateSBM(N, alpha, pi, directed)

}
\references{
[1] Tabouy, P. Barbillon, J. Chiquet. Variationnal inference of Stochastic Block Model from sampled data (2017). arXiv:1707.04141.
}
\seealso{
\code{\link{inferSBM}} and \code{\link{samplingSBM}}
}
