% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSBMcollection-Class.R
\name{smooth}
\alias{smooth}
\title{Smooth path of models in a collection}
\usage{
smooth(Robject, type = c("forward", "backward", "both"),
  split = c("hierarchical", "spectral", "kmeans"),
  control_VEM = list(), cores = 2, iterates = 1, trace = TRUE)
}
\arguments{
\item{Robject}{an object with class missSBM_collection, i.e. an output from \code{\link{estimate}}}

\item{type}{character indicating what kind of ICL smoothing should be use among "forward", "backward" or "both". Default is "foward".}

\item{split}{character indicating the function use to split nodes during the forward algorithm. Either "hierarchical", "spectral" or "kmeans". Default is "hierarchical".}

\item{control_VEM}{a list controlling the variational EM algorithm. See details in \code{\link{estimate}}.}

\item{cores}{integer, the number of cores to use when multiply model are fitted. Default is 2.}

\item{iterates}{integer for the number of iteration in case of foward-backward (aka both) smoothing. Default is 1.}

\item{trace}{logical, control the verbosity. Default to \code{TRUE}.}
}
\value{
an invisible missSBM_collection, in which the ICL has been smoothed
}
\description{
Apply a split and/or merge strategy to the path of model in a collection of SBM, in order to find better initialization. This should result in
a "smoothing" of the ICL, that should be close to concave.
}
