% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class-missSBM_fit.R
\name{summary.missSBM_fit}
\alias{summary.missSBM_fit}
\title{Summary method for a \code{\link{missSBM_fit}}}
\usage{
summary.missSBM_fit(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{missSBM_fit}}}

\item{...}{additional parameters for S3 compatibility.}
}
\value{
a basic printing output
}
\description{
Summary method for a \code{\link{missSBM_fit}}
}
